adding multiple images to demo and extracting the keyboard adapter
