+++ demo/index.js
@@ -1,11 +1,13 @@
 // import TouchEmulator from 'hammer-touchemulator';
 import setupViewer from '../src/index';
 import hammerjsAdapter from '../src/adapters/hammerjs-adapter';
+import keyboardAdapter from '../src/adapters/keyboard-adapter';
 
 // TouchEmulator();
 const canvas = document.getElementById('canvas');
 
 const viewer = setupViewer(canvas, 300, 300);
+keyboardAdapter(canvas, viewer);
 hammerjsAdapter(canvas, viewer);
 viewer
   .addImage('https://images.unsplash.com/photo-1483030096298-4ca126b58199?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&w=1000&q=80')
